+++ program/include/test.hpp
@@ -1,0 +1,9 @@
+//#ifndef TEST_H
+//#define TEST_H
+//
+//
+//#include "prediction.hpp"
+//
+//void test(string expected_series_name, int expected_next_number, vector<int> given);
+//
+//#endif
