+++ program/src/test.cpp
@@ -1,2 +1,7 @@
-#include "prediction.hpp"
-
+//#include "test.hpp"
+//
+//
+//void test(string expected_series_name, int expected_next_number, vector<int> given) {
+//
+//    if (predict(given)->series_name == expected_series_name && predict(given)->next_number == exoected_next_number)
+//}
